% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodonFraction.R
\name{CodonFraction}
\alias{CodonFraction}
\title{Codon Fraction}
\usage{
CodonFraction(seqs, ORF = FALSE, reverseORF = TRUE, label = c())
}
\arguments{
\item{seqs}{is a FASTA file containing nucleotide sequences. The sequences start
with '>'. Also, seqs could be a string vector. Each element of the vector is a nucleotide sequence.}

\item{ORF}{(Open Reading Frame) is a logical parameter. If it is set to true, ORF region of each sequence is considered instead of the original sequence (i.e., 3-frame).}

\item{reverseORF}{is a logical parameter. It is enabled only if ORF is true.
If reverseORF is true, ORF region will be searched in the sequence and also in the reverse complement of the sequence (i.e., 6-frame).}

\item{label}{is an optional parameter. It is a vector whose length is equivalent to the number of sequences. It shows the class of
each entry (i.e., sequence).}
}
\value{
A feature matrix such that the number of columns is 4^3 and the number of rows is equal to the number of sequences.
}
\description{
This function calculates the codon fraction for each sequence.
}
\examples{

fileLNC<-system.file("extdata/Athaliana_LNCRNA.fa",package="ftrCOOL")
mat<-CodonFraction(seqs=fileLNC,ORF=TRUE,reverseORF=FALSE)

}
