% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DisorderB.R
\name{DisorderB}
\alias{DisorderB}
\title{disorder Binary (DisorderB)}
\usage{
DisorderB(
  dirPath,
  binaryType = "strBin",
  outFormat = "mat",
  outputFileDist = ""
)
}
\arguments{
\item{dirPath}{Path of the directory which contains all output files of VSL2. Each file belongs to a sequence.}

\item{binaryType}{It can take any of the following values: ('strBin','logicBin','numBin').
'strBin' (String binary): each dinucleotide is represented by a string containing 2 characters(0-1). order = "10"   disorder="01".
'logicBin' (logical value): Each amino acid is represented by a vector containing 2 logical entries.  order = c(TRUE,FALSE)   disorder=c(FALSE,TRUE).
'numBin' (numeric bin): Each amino acid is represented by a numeric (i.e., integer) vector containing 2 numeric entries.   order = c(1,0)   disorder=c(0,1).}

\item{outFormat}{It can take two values: 'mat' (which stands for matrix) and 'txt'. The default value is 'mat'.}

\item{outputFileDist}{It shows the path and name of the 'txt' output file.}
}
\value{
The output is different depending on the outFormat parameter ('mat' or 'txt').
If outFormat is set to 'mat', it returns a feature matrix for sequences with the same lengths.
The number of rows is equal to the number of sequences and if binaryType is 'strBin', the number of columns is the length of the sequences.
Otherwise, it is equal to (length of the sequences)*2.
If outFormat is 'txt', all binary values will be written to a tab-delimited file. Each line in the file shows the binary format of a sequence.
}
\description{
This function extracts the ordered and disordered amino acids in protein or peptide sequences. The input to the function is provided by VSL2 software. Also, the function
converts order amino acids to '10' and disorder amino acids to '01'.
}
\note{
This function is provided for sequences with the same lengths.
Users can use 'txt' option in outFormat for sequences with different lengths.
Warning: If outFormat is set to 'mat' for sequences with different lengths, it returns an error.
Also, when output format is 'txt', label information is not shown in the text file.
It is noteworthy that 'txt' format is not usable for machine learning purposes if sequences have different sizes. Otherwise 'txt' format
is also usable for machine learning purposes.
}
\examples{

dir = tempdir()
PredDisdir<-system.file("testForder",package="ftrCOOL")

PredDisdir<-paste0(PredDisdir,"/Disdir/")
ad1<-paste0(dir,"/disorderB.txt")

mat<-DisorderB(PredDisdir,binaryType="strBin",outFormat="txt",outputFileDist=ad1)


}
