% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nameKmer.R
\name{nameKmer}
\alias{nameKmer}
\title{nameKmer}
\usage{
nameKmer(k = 3, type = "aa", num = 0)
}
\arguments{
\item{k}{is a numeric value.}

\item{type}{can be one of "aa", "rna", "dna", or "num".}

\item{num}{When type is set to "num", it shows the numeric alphabet( 1,..,,num).}
}
\value{
a string vector of length (20^k for 'aa' type), (4^k for 'dna' type), (4^k for 'rna' type),
and (num^k for 'num' type).
}
\description{
This function creates all possible k-combinations of the given alphabets.
}
\examples{

all_kmersAA<-nameKmer(k=2,type="aa")

all_kmersDNA<-nameKmer(k=3,type="dna")

all_kmersNUM<-nameKmer(k=3,type="num",num=2)
}
