% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/needleman.R
\name{needleman}
\alias{needleman}
\title{Needleman-Wunsch (needleman)}
\usage{
needleman(seq1, seq2, gap = -1, mismatch = -1, match = 1)
}
\arguments{
\item{seq1}{(sequence1) is a string.}

\item{seq2}{(sequence2) is a string.}

\item{gap}{The penalty for gaps in sequence alignment. Usually, it is a negative value.}

\item{mismatch}{The penalty for the mismatch in the sequence alignment. Usually, it is a negative value.}

\item{match}{A score for the match in sequence alignment. Usually, it is a positive value.}
}
\value{
The function returns a number which indicates the similarity between sequence1 and sequence2.
}
\description{
This function works based on Needleman-Wunsch algorithm which computes similarity score of two sequences.
}
\examples{

simScore<-needleman(seq1="Hello",seq2="Hello",gap=-1,mismatch=-2,match=1)
}
\references{
https://gist.github.com/juliuskittler/ed53696ac1e590b413aac2dddf0457f6
}
